update for release
git-svn-id: https://scm.gforge.inria.fr/svn/openalea/branches/release_0_9@2957 04ea39c6-a500-0410-8508-e77fe2a861cf

+++ image/src/image/gui/pixmap_view.py
@@ -202,7 +202,7 @@ class PixmapStackView (PixmapView) :
 		pix = []
 		for z in xrange(data.shape[2]) :
 			#dat = pal[data[:,:,z] ].flatten('F')
-			dat = pal[data[:,:,z] ]
+			dat = pal[ uint32(data[:,:,z]) ]
 			#img = QImage(dat,
 			#             data.shape[0],
 			#             data.shape[1],
